refactor: 기능의 책임자를 변경합니다
