+++ src/main/java/menu/Application.java
@@ -3,7 +3,7 @@ package menu;
 import menu.domain.menu.MenuRecommendService;
 import menu.domain.menu.MenuRepository;
 import menu.ui.view.InputView;
-import menu.ui.view.ResultView;
+import menu.ui.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
@@ -12,7 +12,7 @@ public class Application {
         MenuRecommendService menuRecommendService = new MenuRecommendService(menuRepository);
 
         MenuRecommendController menuRecommendController = new MenuRecommendController(
-                InputView.getInstance(), ResultView.getInstance(), menuRecommendService
+                InputView.getInstance(), OutputView.getInstance(), menuRecommendService
         );
