+++ src/main/java/menu/MenuRecommendController.java
@@ -2,21 +2,21 @@ package menu;
 
 import menu.domain.menu.MenuRecommendService;
 import menu.ui.view.InputView;
-import menu.ui.view.ResultView;
+import menu.ui.view.OutputView;
 
 public class MenuRecommendController {
     private final InputView inputView;
-    private final ResultView resultView;
+    private final OutputView outputView;
     private final MenuRecommendService menuRecommendService;
 
-    public MenuRecommendController(InputView inputView, ResultView resultView, MenuRecommendService menuRecommendService) {
+    public MenuRecommendController(InputView inputView, OutputView outputView, MenuRecommendService menuRecommendService) {
         this.inputView = inputView;
-        this.resultView = resultView;
+        this.outputView = outputView;
         this.menuRecommendService = menuRecommendService;
     }
 
     public void init() {
-        inputView.onBoardingPage();
+        outputView.onBoardingPage();
     }
     public void run() {
     }
