--- src/main/java/menu/ui/view/ResultView.java
@@ -1,11 +1,0 @@
-package menu.ui.view;
-
-public class ResultView {
-
-    private ResultView() {
-
-    }
-    public static ResultView getInstance() {
-        return new ResultView();
-    }
-}
